feat: utilizando state inves de roteamento e corrigindo mudança na renderização ao trocar tabs
