+++ src/pages/login/SignInForm/index.tsx
@@ -1,12 +1,24 @@
 import React from 'react';
 import { useState } from 'react';
-import { Button } from '../Button';
-import { Input } from '../Input';
+import { Button } from '../../../components/Button';
+import { Modal } from '../../../components/Modal';
+import { ResetPassword } from '../../reset_password';
+import { Input } from '../../../components/Input';
 
-export const SignUpForm = () => {
+// todo import api from '../../lib/api';
+
+export const SignInForm = () => {
+  // constantes que controlam a abertura e fechamento do modal
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const handleOpenModal = () => {
+    setIsModalOpen(true);
+  };
+  const handleCloseModal = () => {
+    setIsModalOpen(false);
+  };
+
   // constantes que controlam os dados do formulário
   const [formData, setFormData] = useState({
-    name: '',
     email: '',
     password: '',
   });
@@ -19,26 +31,12 @@ export const SignUpForm = () => {
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
     console.log(formData);
+    // todo api.post('/login', formData);
   };
 
   return (
     <form id="form" className="flex flex-col space-y-6" onSubmit={handleSubmit}>
       <div className="space-y-6">
-        <div className="flex flex-col space-y-2">
-          <label htmlFor="name_input" className="text-cream-100 text-xs w-max">
-            Nome
-          </label>
-          <Input
-            id="name_input"
-            name="name"
-            type="text"
-            value={formData.name}
-            onChange={handleChange}
-            required
-            placeholder="Digite seu nome"
-          ></Input>
-        </div>
-
         <div className="flex flex-col space-y-2">
           <label htmlFor="email_input" className="text-cream-100 text-xs w-max">
             Email
@@ -49,7 +47,6 @@ export const SignUpForm = () => {
             type="email"
             value={formData.email}
             onChange={handleChange}
-            required
             placeholder="Digite seu email"
           ></Input>
         </div>
@@ -67,15 +64,26 @@ export const SignUpForm = () => {
             type="password"
             value={formData.password}
             onChange={handleChange}
-            required
             placeholder="Digite sua senha"
           ></Input>
+          <button
+            className="flex text-xs underline text-cream-100"
+            onClick={handleOpenModal}
+          >
+            Esqueci a senha
+          </button>
+
+          {isModalOpen && (
+            <Modal title="Redefinição de senha" closeModal={handleCloseModal}>
+              <ResetPassword />
+            </Modal>
+          )}
         </div>
       </div>
 
       <div className="flex justify-center">
         <Button
-          label="Inscreva-se"
+          label="Entrar"
           variant="primary"
           className="w-max"
           type="submit"
